requirements.txt
future
isort!=4.3.0
pip
psutil>=3.1.0
pycparser
pyelftools
python-ptrace>=0.8
ROPgadget
six
unicorn>=1.0.0
pygments
https://github.com/aquynh/capstone/archive/next.zip#subdirectory=bindings/python
enum34
pytest
